added ClearTable() test
